// DutchTreat_th/client/src/app/shared/Product.ts
export class Product {
         id: number=0;
        category: string="";
        size: string="";
        price: number=0;
        title: string="";
        artDescription: string="";
        artDating: string="";
        artId: string="";
    artist: string = "";
    artistBirthDate: Date = new Date;
    artistDeathDate: Date = new Date;
    artistNationality: string="";
    //constructor() {
    //    this.id = 0;
    //    this.category = "";
    //    this.size = "";
    //    this.price = 0;
    //    this.title = "";
    //    this.artDescription = "";
    //    this.artDating = "";
    //    this.artId = "";
    //    this.artist = "";
    //    this.artistBirthDate = new Date;
    //    this.artistDeathDate = new Date;
    //    this.artistNationality = "";
    //}
}